+++ CHANGELOG.md
@@ -1,8 +1,12 @@
 # Unreleased
 
-- **BREAKING** [PENDING] Update dependency `vulkano` & `vulkano-shaders` to `0.29`.
+- **BREAKING** Update dependency `vulkano` & `vulkano-shaders` to `0.29`.
 - **BREAKING** `Basalt::swap_caps()` has been replaced by `Basalt::surface_capabilities`, `Basalt::surface_formats`, & `Basalt::surface_present_modes`.
 - **BREAKING** `Basalt::current_extent()` now takes `FullScreenExclusive` as an argument.
+- **BREAKING** Renamed `Qwery` to `Qwerty`...
+- **BREAKING** `BinColor::as_tuple()` has been removed and replaced by `BinColor::as_array()`.
+- **BREAKING** `atlas::Coords` methods `top_left`, `top_right`, `bottom_left`, & `bottom_right` now return `[f32; 2]` instead of `(f32, f32)`.
+
 - **BEHAVIOR** `Camera::mouse_inside()` will now return false for `Bin`'s that have `pass_events` set to `Some(true)`.
 - Fixed various instances of buffers being created with zero size.
 - Fixed wrong image values for Atlas's empty image.